+++ src/tiled/mainwindow.cpp
@@ -1009,6 +1009,15 @@ void MainWindow::dropEvent(QDropEvent *e)
 
 void MainWindow::resizeEvent(QResizeEvent *e)
 {
+    // When the window is maximized, we need to delay restoring the internal
+    // layout until after the second resize event (issue #590). This does not
+    // appear to affect macOS, where only a single resize event is observed.
+    if (!mHasRestoredLayout)
+#ifndef Q_OS_MAC
+        if (!isMaximized() || e->oldSize().isValid())
+#endif
+            restoreLayout();
+
     if (mPopupWidget)
         updatePopupGeometry(e->size());
 }
@@ -2304,13 +2313,24 @@ void MainWindow::readSettings()
         restoreGeometry(geom);
     else
         resize(Utils::dpiScaled(QSize(1200, 700)));
-    restoreState(preferences::mainWindowState);
+
+    // Make sure layout is restored eventually (see resizeEvent)
+    QTimer::singleShot(200, this, &MainWindow::restoreLayout);
+
     updateRecentFilesMenu();
     updateRecentProjectsMenu();
 
-    mDocumentManager->restoreState();
-
     mLockLayout->setChecked(preferences::mainWindowLocked);
+}
+
+void MainWindow::restoreLayout()
+{
+    if (mHasRestoredLayout)
+        return;
+
+    mHasRestoredLayout = true;
+    restoreState(preferences::mainWindowState);
+    mDocumentManager->restoreState();
 }
 
 void MainWindow::updateWindowTitle()